rename: 버튼 파일 컴포넌트 파일로 이동
