+++ src/components/common/Button/ToggleButton.styled.js
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import colors from '../../styles/colors';
+import colors from '../../../styles/colors';
 
 export const ToggleDiv = styled.div`
   width: 12.2rem;
